+++ CHANGELOG.md
@@ -4,6 +4,27 @@ GovReady-Q Release Notes
 v999 (February XX, 2021)
 ------------------------
 
+**Feature changes**
+
+* Remove a component and its statements from a system.
+
+**UI changes**
+
+* Add a button to system selected component page to remove a component from the system.
+* Moved "Add a component" to a system drop down to top of selected component page.
+
+**Developer changes**
+
+* Move inclusion of `edit-component-modal.html` from `base.html` to `components/element_detail_tabs.html`.
+* Fix sort control order in `component_library_component` on the `components/element_detail_tabs.html` using the `natsort` package to sort SID correctly.
+
+**Data changes**
+
+* Alter Element description field to be blank and none.
+
+v0.9.1.52 (February 16, 2021)
+-----------------------------
+
 Add System Assessment Report tracking to associate assessments and evidence with the system.
 Add initial dynamic status information to the project page.
 Project page displays mini-dashboard of compliance stats.